requirements.txt
# Core ML & NLP
transformers>=4.41.0
datasets>=2.19.0
huggingface_hub>=0.23.0
accelerate>=0.28.0
evaluate>=0.4.0

# Deep-learning runtime
torch>=2.2.0     # install the CUDA build separately if using GPU

# LLM providers & tooling
openai>=1.30.0
google-generativeai>=0.5.0
google-api-core>=2.18.0
anthropic>=0.25.0
tiktoken>=0.7.0
tenacity>=8.3.0
pydantic>=2.7.0

# Hyper-parameter tuning
ray[tune]>=2.20.0

# Data science stack
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0

# Misc
python-dotenv>=1.0.1
protobuf>=4.25.0 